mpi/abort: add CVAR to disable Abort message
Sometimes an application non-collectively encounters a condition in
which in must exit (usually with a message to the end user), but that is
not a "crash".  This results in

  application called MPI_Abort(MPI_COMM_WORLD, 1) - process 0

and perhaps extra cruft from the process manager.  These lines obscure
any information the application wrote for the end user, causing users to
stop reading or otherwise get confused.  This commit makes it possible
for the application to suppress the line above, though it does not
suppress output from the process manager.

The CVAR is normally read at load-time, but can be modified using the
MPI_T_ interface.  For example, a quiet error handler could be written

  void exit_quietly(int retcode) {
    int tsupport, num, i;
    MPI_T_init_thread(MPI_THREAD_SINGLE,&tsupport);
    MPI_T_cvar_get_num(&num);
    for (i=0; i<num; i++) {
      char name[100];
      int namelen = sizeof(name), verbose, bind, scope;
      MPI_Datatype datatype;
      MPI_T_cvar_get_info(i, name, &namelen, &verbose, &datatype, NULL, NULL, NULL, &bind, &scope);
      if (!strcmp(name,"MPIR_CVAR_SUPPRESS_ABORT_MESSAGE")) {
        MPI_Comm comm;
        MPI_T_cvar_handle handle;
        int count, one = 1;
        MPI_T_cvar_handle_alloc(i, &comm, &handle, &count);
        if (datatype == MPI_INT && count == 1) MPI_T_cvar_write(handle, &one);
        MPI_T_cvar_handle_free(&handle);
      }
    }
    MPI_T_finalize();
    MPI_Abort(MPI_COMM_WORLD, retcode);
  }

Reported-by: Barry Smith <bsmith@mcs.anl.gov>
Based-on-patch-by: Jeff Hammond <jeff.science@gmail.com>