address FIXME w.r.t. atomic win_ptr->my_counter
As noted in the comments, my_counter needs atomic accesses to be
thread-safe.  This commit adds all such access.

The changes are protected with a new preprocessor macro since the
overhead of atomics might be high and need to be disabled for
non-MULTIPLE usage.

After making these changes, it strikes me that a macro for my_counter
operations would make the code cleaner, e.g.

    OPA_store_int(&(win_ptr->my_counter), 0);
    win_ptr->my_counter = 0;

could instead be

    RMA_COUNTER_STORE((win_ptr->my_counter),0);

with an O(1) change of

    #define RMA_COUNTER_STORE(ctr,val) OPA_store_int(&(ctr), val);
    #define RMA_COUNTER_STORE(ctr,val) {ctr = val};

or something like this...